+++ src/page/index.tsx
@@ -1,5 +1,6 @@
 import Dashboard from "./dashboard/dashboard";
 import Intro from "./intro/intro";
 import Layout from "./layout/layout";
+import ProjectsPage from "../page/dashboard/project";
 
-export { Dashboard, Intro, Layout };
+export { Dashboard, Intro, Layout, ProjectsPage };
