+++ src/router/router.tsx
@@ -1,5 +1,5 @@
 import { Routes, Route, BrowserRouter } from "react-router-dom";
-import { Dashboard, Intro, Layout } from "../page/index";
+import { Dashboard, Intro, Layout,ProjectsPage } from "../page/index";
 
 const AppRouter = () => {
   return (
@@ -8,7 +8,7 @@ const AppRouter = () => {
         <Route path="/intro" element={<Intro />} />
         <Route path="/" element={<Layout />}>
           <Route index element={<Dashboard />} />
-          <Route path="project" element={<Dashboard />} />
+          <Route path="project" element={<ProjectsPage />} />
           <Route path="sprint" element={<Dashboard />} />
           <Route path="team" element={<Dashboard />} />
           <Route path="reports" element={<Dashboard />} />
