--- src/styles/styled.d.ts
@@ -1,8 +1,0 @@
-import "styled-components";
-import { lightTheme } from "./theme";
-
-type Theme = typeof lightTheme;
-
-declare module "styled-components" {
-  export interface DefaultTheme extends Theme {}
-}
